+++ packages/core/src/utils/index.ts
@@ -49,3 +49,5 @@ export { genShortID } from './shortId';
 export { getBundleId } from './sourceMaps';
 
 export { dateNow } from './date';
+
+export { getCircularDependencyReplacer, stringifyExternalJson } from './json';