+++ packages/web-sdk/src/instrumentations/console/instrumentation.ts
@@ -1,4 +1,14 @@
-import { allLogLevels, BaseInstrumentation, defaultLogArgsSerializer, LogLevel, VERSION } from '@grafana/faro-core';
+import {
+  allLogLevels,
+  BaseInstrumentation,
+  defaultErrorArgsSerializer,
+  defaultLogArgsSerializer,
+  LogArgsSerializer,
+  LogLevel,
+  VERSION,
+} from '@grafana/faro-core';
+
+import { getDetailsFromConsoleErrorArgs } from '../errors/getErrorDetails';
 
 import type { ConsoleInstrumentationOptions } from './types';
 
@@ -7,14 +17,20 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
   readonly version = VERSION;
 
   static defaultDisabledLevels: LogLevel[] = [LogLevel.DEBUG, LogLevel.TRACE, LogLevel.LOG];
+  static consoleErrorPrefix = 'console.error: ';
+  private errorSerializer: LogArgsSerializer = defaultLogArgsSerializer;
 
   constructor(private options: ConsoleInstrumentationOptions = {}) {
     super();
   }
 
   initialize() {
-    this.logDebug('Initializing\n', this.options);
     this.options = { ...this.options, ...this.config.consoleInstrumentation };
+
+    const serializeErrors = this.options?.serializeErrors || !!this.options?.errorSerializer;
+    this.errorSerializer = serializeErrors
+      ? (this.options?.errorSerializer ?? defaultErrorArgsSerializer)
+      : defaultLogArgsSerializer;
 
     allLogLevels
       .filter(
@@ -25,7 +41,25 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
         console[level] = (...args) => {
           try {
             if (level === LogLevel.ERROR && !this.options?.consoleErrorAsLog) {
-              this.api.pushError(new Error('console.error: ' + defaultLogArgsSerializer(args)));
+              const { value, type, stackFrames } = getDetailsFromConsoleErrorArgs(args, this.errorSerializer);
+
+              if (value && !type && !stackFrames) {
+                this.api.pushError(new Error(ConsoleInstrumentation.consoleErrorPrefix + value));
+                return;
+              }
+
+              this.api.pushError(new Error(ConsoleInstrumentation.consoleErrorPrefix + value), { type, stackFrames });
+            } else if (level === LogLevel.ERROR && this.options?.consoleErrorAsLog) {
+              const { value, type, stackFrames } = getDetailsFromConsoleErrorArgs(args, this.errorSerializer);
+
+              this.api.pushLog(value ? [ConsoleInstrumentation.consoleErrorPrefix + value] : args, {
+                level,
+                context: {
+                  value: value ?? '',
+                  type: type ?? '',
+                  stackFrames: stackFrames?.length ? defaultErrorArgsSerializer(stackFrames) : '',
+                },
+              });
             } else {
               this.api.pushLog(args, { level });
             }