+++ packages/web-sdk/src/utils/index.ts
@@ -10,6 +10,4 @@ export {
 
 export { throttle } from './throttle';
 
-export { getCircularDependencyReplacer, stringifyExternalJson } from './json';
-
 export { getIgnoreUrls } from './url';